#5 finishing off poller (first complete version) and fixed scrapers
